[CHORE] 코드리뷰 반영 - BaseViewController에서 빼서 Struct화 시키기 (#345)
