+++ Manito/Manito/Global/Base/BaseViewController.swift
@@ -146,39 +146,6 @@ class BaseViewController: UIViewController {
     private func setupInteractivePopGestureRecognizer() {
         self.navigationController?.interactivePopGestureRecognizer?.delegate = self
     }
-    
-    func showToast(message: String) {
-        let toastLabel = UILabel()
-        toastLabel.backgroundColor = .grey001
-        toastLabel.textColor = .black
-        toastLabel.font = .font(.regular, ofSize: 14)
-        toastLabel.textAlignment = .center
-        toastLabel.text = message
-        toastLabel.alpha = 0
-        toastLabel.layer.cornerRadius = 10
-        toastLabel.clipsToBounds = true
-        self.view.addSubview(toastLabel)
-        toastLabel.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.bottom.equalToSuperview().inset(150)
-            $0.width.equalTo(140)
-            $0.height.equalTo(40)
-        }
-        UIView.animate(withDuration: 0.7, animations: {
-            toastLabel.alpha = 0.8
-        }, completion: { isCompleted in
-            UIView.animate(withDuration: 0.7, delay: 0.5, animations: {
-                toastLabel.alpha = 0
-            }, completion: { isCompleted in
-                toastLabel.removeFromSuperview()
-            })
-        })
-    }
-    
-    func touchUpToShowToast(code: String) {
-        UIPasteboard.general.string = code
-        self.showToast(message: TextLiteral.detailWaitViewControllerCopyCode)
-    }
 }
 
 extension BaseViewController: UIGestureRecognizerDelegate {
